move scraping to separate file
